Remove FastMath
hotspot has intrinsics for almost everything we use it for

+++ common/api/src/main/java/com/dfsek/terra/api/util/collection/ProbabilityCollection.java
@@ -51,25 +51,25 @@ public class ProbabilityCollection<E> implements Collection<E> {
     @SuppressWarnings("unchecked")
     public E get(NoiseSampler n, double x, double y, double z, long seed) {
         if(array.length == 0) return null;
-        return (E) array[MathUtil.normalizeIndex(n.noise(seed, x, y, z), array.length)];
+        return (E) array[(int) MathUtil.normalizeIndex(n.noise(seed, x, y, z), array.length)];
     }
     
     @SuppressWarnings("unchecked")
     public E get(NoiseSampler n, Vector3Int vector3Int, long seed) {
         if(array.length == 0) return null;
-        return (E) array[MathUtil.normalizeIndex(n.noise(seed, vector3Int.getX(), vector3Int.getY(), vector3Int.getZ()), array.length)];
+        return (E) array[(int) MathUtil.normalizeIndex(n.noise(seed, vector3Int.getX(), vector3Int.getY(), vector3Int.getZ()), array.length)];
     }
     
     @SuppressWarnings("unchecked")
     public E get(NoiseSampler n, Vector3 vector3Int, long seed) {
         if(array.length == 0) return null;
-        return (E) array[MathUtil.normalizeIndex(n.noise(seed, vector3Int.getX(), vector3Int.getY(), vector3Int.getZ()), array.length)];
+        return (E) array[(int) MathUtil.normalizeIndex(n.noise(seed, vector3Int.getX(), vector3Int.getY(), vector3Int.getZ()), array.length)];
     }
     
     @SuppressWarnings("unchecked")
     public E get(NoiseSampler n, double x, double z, long seed) {
         if(array.length == 0) return null;
-        return (E) array[MathUtil.normalizeIndex(n.noise(seed, x, z), array.length)];
+        return (E) array[(int) MathUtil.normalizeIndex(n.noise(seed, x, z), array.length)];
     }
     
     @SuppressWarnings("unchecked")
